Refactor: viewmodel, screen 폴더 생성
