+++ feature-perfume/src/main/java/com/hmoa/feature_perfume/viewmodel/SpecificCommentViewmodel.kt
@@ -1,4 +1,4 @@
-package com.hmoa.feature_perfume
+package com.hmoa.feature_perfume.viewmodel
 
 import androidx.lifecycle.ViewModel
 import com.hmoa.core_domain.repository.PerfumeCommentRepository
